+++ internal/pkg/codehub/repository/postgres/repository.go
@@ -14,15 +14,15 @@ type Repo struct {
 func NewRepository(db *sqlx.DB) Repo {
 	return Repo{DB: db}
 }
-func (R *Repo) AddStar(userID int, repoID int) error {
+func (R *Repo) AddStar(userID int64, repoID int64) error {
 	return nil
 }
 
-func (R *Repo) DelStar(userID int, repoID int) error {
+func (R *Repo) DelStar(userID int64, repoID int64) error {
 	return nil
 }
 
-func (R *Repo) GetStarredRepo(userID int) ([]gitmodels.Repository, error) {
+func (R *Repo) GetStarredRepo(userID int64) ([]gitmodels.Repository, error) {
 	return []gitmodels.Repository{}, nil
 }
 
@@ -34,22 +34,22 @@ func (R *Repo) UpdateIssue(issue models.Issue) error {
 	return nil
 }
 
-func (R *Repo) CloseIssue(issueID int) error {
+func (R *Repo) CloseIssue(issueID int64) error {
 	return nil
 }
 
-func (R *Repo) GetIssuesList(repoID int) ([]models.Issue, error) {
+func (R *Repo) GetIssuesList(repoID int64) ([]models.Issue, error) {
 	return []models.Issue{}, nil
 }
 
-func (R *Repo) CheckAccessIssue(userID, issueID int) (perm.Permission, error) {
+func (R *Repo) CheckAccessIssue(userID, issueID int64) (perm.Permission, error) {
 	return perm.AdminAccess(), nil
 }
 
-func (R *Repo) CheckAccessRepo(userID, repoID int) (perm.Permission, error) {
+func (R *Repo) CheckAccessRepo(userID, repoID int64) (perm.Permission, error) {
 	return perm.AdminAccess(), nil
 }
 
-func (R *Repo) GetIssues(issueID int) (models.Issue, error) {
+func (R *Repo) GetIssues(issueID int64) (models.Issue, error) {
 	return models.Issue{}, nil
 }
